ADD: Create a conditional to view the waiter or cook pages
